Part way through refactoring the safeguarding section to it's own controller. Added remaining form elements. Changed the naming convention for getting translations, will improve this further later when I decide best place to put this instead of the component. probably the controller (look at iOS and other for good practice) Started Javascript component for showing/hiding based on radio value.
