+++ src/AppBundle/Controller/ReportController.php
@@ -275,44 +275,6 @@ class ReportController extends Controller
         ];
     }
 
-    /**
-     * @Route("/report/{reportId}/safeguarding", name="safeguarding")
-     * @Template()
-     */
-    public function safeguardingAction($reportId)
-    {
-        $util = $this->get('util');
-        $report = $util->getReport($reportId);
-        $request = $this->getRequest();
-        
-        $form = $this->createForm(new FormDir\SafeguardingType(), $report);
-        
-
-        if($request->getMethod() == 'POST'){
-            $form->handleRequest($request);
-            
-            if($form->get('save')->isClicked()){
-                if($form->isValid()){
-                    $data = $form->getData();
-                    $data->keepOnlyRelevantSafeguardingData();
-
-                    $this->get('apiclient')->putC('report/' .  $report->getId(),$data);
-                    $translator = $this->get('translator');
-
-                    $this->get('session')->getFlashBag()->add('action', 'page.safeguardinfoSaved');
-
-                    return $this->redirect($this->generateUrl('report_safeguarding', ['reportId'=>$reportId]));
-                }
-            }
-        }
-
-        return[ 'report' => $report,
-                'client' => $util->getClient($report->getClient()),
-                'form' => $form->createView()
-              ];
-    }
-    
-
     private function groupAssets($assets)
     {
         $assetGroups = array();
