+++ src/AppBundle/Resources/assets/scss/components/_help.scss
@@ -16,8 +16,7 @@
 }
 
 .report-declaration #contact-block,
-.report-money-input #contact-block,
-.safeguarding-page #contact-block 
+.report-money-input #contact-block
 {
     display: none;
     @include media(mobile) {
